requirements.txt
requests
python-dotenv
pandas
scikit-learn
fastapi
uvicorn
sqlalchemy
psycopg2-binary